backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-ckeditor==6.2.0
django-jazzmin==2.4.9
django-js-asset==2.0.0
django-mptt==0.13.4
django-parler==2.3
django-rosetta==0.9.8
idna==3.3
Pillow==9.0.1
polib==1.1.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
